world load progress check

+++ src/assembler/events.rs
@@ -4,4 +4,4 @@ use bevy::prelude::*;
 pub struct PrepareLevelEvent(pub String);
 
 #[derive(Event)]
-pub struct PostProgresssEvent(pub String, pub usize, pub usize);
+pub struct PostProgresssEvent(pub String, pub u32, pub u32);
